Added market stats, historical data, event system, better deserialization & PartialBookDepth support

+++ BinanceAPI.NET/Core/Models/Objects/Entities/BinanceKlineCandlestick.cs
@@ -26,25 +26,25 @@ namespace BinanceAPI.NET.Core.Models.Objects.Entities
         [JsonProperty("L")]
         public long LastTradeId { get; set; }
         [JsonProperty("o")]
-        public double OpenPrice { get; set; }
+        public decimal OpenPrice { get; set; }
         [JsonProperty("c")]
-        public double ClosePrice { get; set; }
+        public decimal ClosePrice { get; set; }
         [JsonProperty("h")]
-        public double HighPrice { get; set; }
+        public decimal HighPrice { get; set; }
         [JsonProperty("l")]
-        public double LowPrice { get; set; }
+        public decimal LowPrice { get; set; }
         [JsonProperty("v")]
-        public double BaseAssetVolume { get; set; }
+        public decimal BaseAssetVolume { get; set; }
         [JsonProperty("n")]
         public int NumberOfTrades { get; set; }
         [JsonProperty("x")]
         public bool IsClosed { get; set; }
         [JsonProperty("q")]
-        public double QuoteAssetVolume { get; set; }
+        public decimal QuoteAssetVolume { get; set; }
         [JsonProperty("V")]
-        public double TakerBuyBaseVolume { get; set; }
+        public decimal TakerBuyBaseVolume { get; set; }
         [JsonProperty("Q")]
-        public double TakerBuyQuoteVolume { get; set; }
+        public decimal TakerBuyQuoteVolume { get; set; }
         [JsonProperty("B")]
         public string ignore { get; set; }
